refactor(csa-server-workers/tests): readLineFromWebSocket を ws_test_helpers.ts に共有化 (#699 follow-up)
PR #699 (issue #622 PR3a) merge 後、claude[bot] review P2 で指摘された
`tests/miniflare_smoke/lobby.test.ts` と `tests/miniflare_smoke/rate_limit.test.ts`
間の `readLineFromWebSocket` 重複を解消する。

## Why

両 inline 版は実装が完全に一致 (formatting のみ差異) で、将来 timeout 既定値や
close handler 挙動を変更したいときに両ファイルを同時に修正する必要があり、
維持性が悪い。テスト専用 helper として切り出すことで:

- 仕様変更の単一窓口化 (例: timeout 既定値の調整、エラー文字列フォーマット)
- 第 3 の smoke (`game_room` 直接通電 test 等、将来追加見込み) からの再利用が
  低コストで可能に

## What

新設 `tests/miniflare_smoke/ws_test_helpers.ts`:
- `WebSocketLineBuffer` interface (略語回避、`WsLineBuffer` ではなく)
- `readLineFromWebSocket(ws): WebSocketLineBuffer` 関数
- 旧 inline 2 版と **bytewise-identical** な挙動 (3000 ms 既定 timeout、
  `"connection closed"` / `"takeLine timeout after Xms"` エラー文言、queue
  優先 / `waiters.splice` cleanup の lifecycle 全て一致)
- module / 関数内コメントは **WHY 中心** (`CLAUDE.md` 規約準拠):
  - 3000 ms 既定値の根拠 (smoke green の観測実績を不変条件として固定)
  - listener 二重登録の警告 (同 ws で複数 buffer NG)
  - `queue` + `waiters` split 設計の race-free 保証
  - close 経路の `waiters` reject の理由 (vitest test hang 回避)
  - `setTimeout` cleanup の二重解決防止意図

`tests/miniflare_smoke/lobby.test.ts`:
- inline `interface LobbyLineBuffer` + `function readLineFromWebSocket` 削除
- import `readLineFromWebSocket` from `./ws_test_helpers` 追加

`tests/miniflare_smoke/rate_limit.test.ts`:
- inline `function readLineFromWebSocket` (`harness 側に出すと scope が
  広すぎる` という当時のコメント含む) 削除
- import `readLineFromWebSocket` from `./ws_test_helpers` 追加

## Test

- `cargo fmt -p rshogi-csa-server-workers && cargo clippy -p rshogi-csa-server-workers
  --tests --all-targets -- -D warnings` clean
- `cargo test -p rshogi-csa-server-workers` 305 lib + 35 integration tests 全 pass
- `pnpm run test:smoke` 11 files / 34 tests 全 pass
  (rate_limit.test.ts 8 / lobby.test.ts 6 含む、共有化後も挙動不変を実証)

## Codex local review

APPROVE_WITH_NITS (1st: comment WHY 化 / `WsLineBuffer` → `WebSocketLineBuffer`
リネーム指摘) → 反映後 APPROVE (2nd) で push 前 review 完了。

Refs: #622 #699
